Add botão imprimir tela principal
