+++ src/@legacy/paginas/NAAPA/Encaminhamento/Cadastro/cadastroEncaminhamentoNAAPABotoesAcao.js
@@ -1,9 +1,9 @@
 /* eslint-disable react/prop-types */
 import { Col, Row } from 'antd';
-import React, { useEffect } from 'react';
+import React, { useEffect} from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useRouteMatch } from 'react-router-dom';
-import { Button, Colors } from '~/componentes';
+import { Button, Colors,Loader } from '~/componentes';
 import BotaoVoltarPadrao from '~/componentes-sgp/BotoesAcaoPadrao/botaoVoltarPadrao';
 import {
   SGP_BUTTON_ALTERAR_CADASTRAR,
@@ -11,6 +11,7 @@ import {
   SGP_BUTTON_ENCERRAR_ENCAMINHAMENTO_NAAPA,
   SGP_BUTTON_PROXIMO_PASSO,
   SGP_BUTTON_SALVAR_RASCUNHO,
+  SGP_BUTTON_IMPRIMIR,
 } from '~/constantes/ids/button';
 import {
   history,
@@ -29,13 +30,13 @@ import {
 } from '~/redux/modulos/encaminhamentoNAAPA/actions';
 import QuestionarioDinamicoFuncoes from '~/componentes-sgp/QuestionarioDinamico/Funcoes/QuestionarioDinamicoFuncoes';
 import situacaoNAAPA from '~/dtos/situacaoNAAPA';
+import { MENSAGEM_SOLICITACAO_RELATORIO_SUCESSO } from '~/constantes';
 
 const CadastroEncaminhamentoNAAPABotoesAcao = props => {
-  const { mostrarBusca, setMostrarBusca } = props;
+  const { mostrarBusca, setMostrarBusca,idEncaminhamento } = props;
 
   const routeMatch = useRouteMatch();
   const dispatch = useDispatch();
-
   const aluno = useSelector(state => state.localizarEstudante.aluno);
 
   const usuario = useSelector(state => state.usuario);
@@ -135,7 +136,16 @@ const CadastroEncaminhamentoNAAPABotoesAcao = props => {
       }
     }
   };
+  const onClickImpressao = async () => {
+    const resultado = await ServicoNAAPA.imprimir(
+      [idEncaminhamento]
+    )
+      .catch(e => erros(e))
+      .finally();
 
+      sucesso(MENSAGEM_SOLICITACAO_RELATORIO_SUCESSO);
+    
+  };
   const onClickSalvarRascunho = async () => {
     const resposta = await ServicoNAAPA.salvarPadrao(
       encaminhamentoId,
@@ -201,7 +211,16 @@ const CadastroEncaminhamentoNAAPABotoesAcao = props => {
       <Col>
         <BotaoVoltarPadrao onClick={() => onClickVoltar()} />
       </Col>
-
+      <Col>
+        <Button
+          bold
+          border
+          icon='print'
+          color={Colors.Azul}
+          id={SGP_BUTTON_IMPRIMIR}
+          onClick={onClickImpressao}
+        />
+      </Col>
       {mostrarBusca ? (
         <Col>
           <Button
